group post comments this will add the post user has commented on however the like and comment count relate to original group post and not the user's comment, since those information for comments are not exposed by linkedin API
Comment Output Fields

Parameter Definition
id Default. The comment ID.
text The comment text. Character limit?
creator The name of the person who created the comment. See
Profile Fieldsto view the person fields available.
creation-timestamp The timestamp for when the comment was created.
relation-to-viewer:(available-actions) Valid values are:
flag-as-inappropriate
delete